refactor(packages): consolidate shared internals
- merge Cursor native exec response handling and flatten provider barrels
- scope Codex fetch failure tracking to adapter instances
- share transport mechanics and canonical package-local types
