--- packages/core-ai-gateway/src/codex/index.ts
@@ -1,2 +1,0 @@
-export * from "./responses/index.js";
-export * from "./credentials.js";
