--- packages/core-ai-gateway/src/cursor/index.ts
@@ -1,2 +1,0 @@
-export * from "./native/index.js";
-export * from "./credentials.js";
