+++ packages/core-ai-gateway/src/index.ts
@@ -30,7 +30,8 @@ export * from "./kimi/quota.js";
 export * from "./opencode-go/quota.js";
 export * from "./opencode-go/usage-scan.js";
 export * from "./canonical/index.js";
-export * from "./cursor/index.js";
+export * from "./cursor/credentials.js";
+export * from "./cursor/native/index.js";
 export * from "./cursor/diagnostic-log.js";
 export * from "./gateway-router/router.js";
 export * from "./gateway-router/types.js";
@@ -48,6 +49,7 @@ export * from "./models.js";
 export * from "./settings/index.js";
 export * from "./settings/store.js";
 export * from "./opencode-go/index.js";
-export * from "./codex/index.js";
+export * from "./codex/credentials.js";
+export * from "./codex/responses/index.js";
 export * from "./kimi/index.js";
 export * from "./transport/sse-keepalive.js";
